requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.24.3
scipy==1.11.3
scikit-image==0.21.0
Pillow==10.1.0